+++ html_form_builder/doc/changelog.rst
@@ -1,3 +1,15 @@
+v1.8.7
+======
+* Embed forms was missing the anti bot measure(my_pie)
+
+v1.8.6
+======
+* Binary file type
+
+v1.8.5
+======
+* Checkbox Group snippet
+
 v1.8.4
 ======
 * Input group snippet