+++ html_form_builder_snippets/doc/changelog.rst
@@ -1,3 +1,17 @@
+v.1.8.12
+========
+* Add integer and selection type fields to input group
+
+v.1.8.11
+========
+* Binary snippet
+
+v.1.8.10
+========
+* Checkbox Group snippet (many2many)
+* Texboxes can now use integer fields
+* Textareas can now use char fields
+
 v.1.8.9
 =======
 * Input group snippet (one2many)